Implement new route for managers to check last purchases, fix some ticket- and event-related bugs

+++ backend/app/api/routes/tickets.py
@@ -1,14 +1,61 @@
 from collections.abc import Sequence
 from uuid import UUID
 
-from fastapi import APIRouter, HTTPException
+from fastapi import APIRouter, HTTPException, status
 from sqlmodel import select
 
+from app import crud
 from app.api.deps import CurrentUser, SessionDep
 from app.api.websockets import manager
-from app.models import Event, EventPublic, Ticket, Voucher
+from app.models import Event, EventPublic, Role, Ticket, TicketPurchaseResponse, Voucher
 
 router = APIRouter()
+
+
+@router.get("/my-tickets", response_model=Sequence[Ticket])
+def list_my_tickets(
+    *, session: SessionDep, current_user: CurrentUser
+) -> Sequence[Ticket]:
+    """
+    List all tickets purchased by the current user.
+    """
+
+    tickets = session.exec(
+        select(Ticket).where(Ticket.user_id == current_user.id)
+    ).all()
+    return tickets
+
+
+@router.get("/activities", response_model=Sequence[TicketPurchaseResponse])
+def read_manager_ticket_purchases(
+    session: SessionDep, current_user: CurrentUser, limit: int = 10
+) -> Sequence[TicketPurchaseResponse]:
+    """
+    Get the last X ticket purchases.
+    """
+
+    if current_user.role != Role.EVENTMANAGER:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Only event managers can access ticket purchases",
+        )
+
+    results = crud.get_manager_ticket_purchases(session, current_user.id, limit)
+
+    ticket_purchases = [
+        TicketPurchaseResponse(
+            ticket_id=ticket.ticket_id,
+            event_id=event.id,
+            event_title=event.title,
+            user_id=user.id,
+            user_email=user.email,
+            quantity=ticket.quantity,
+            purchase_date=ticket.purchase_date,
+        )
+        for ticket, event, user in results
+    ]
+
+    return ticket_purchases
 
 
 @router.post("/{id}/buy", response_model=EventPublic)
@@ -49,24 +96,11 @@ async def buy_ticket(
     event.sold_tickets += quantity
     ticket = Ticket(event_id=event_id, user_id=current_user.id, quantity=quantity)
     session.add(ticket)
-    session.delete(voucher)
+    if voucher_id:
+        session.delete(voucher)
     session.commit()
     session.refresh(event)
     await manager.broadcast(
         {"type": "ticket_purchase", "quantity": quantity, "event": event.model_dump()}
     )
     return event
-
-
-@router.get("/my-tickets", response_model=Sequence[Ticket])
-def list_my_tickets(
-    *, session: SessionDep, current_user: CurrentUser
-) -> Sequence[Ticket]:
-    """
-    List all tickets purchased by the current user.
-    """
-
-    tickets = session.exec(
-        select(Ticket).where(Ticket.user_id == current_user.id)
-    ).all()
-    return tickets
